backend/src/requirements.txt
pyquery
requests==2.7.0
pymysql
flask-cors
python-dateutil
gunicorn
celery